refactor: remove ForgeExecutorService and integrate chat handling directly in ForgeAPI
