+++ crates/forge_api/src/lib.rs
@@ -1,4 +1,3 @@
-mod executor;
 mod forge_api;
 
 pub use forge_api::*;
